Bottom sheet for transaction implemented
