--- lib/widgets/user_transaction.dart
@@ -1,48 +1,0 @@
-import 'package:flutter/material.dart';
-import '../models/transaction.dart';
-import '../widgets/new_transaction.dart';
-import '../widgets/transaction_list.dart';
-
-class UserTransactions extends StatefulWidget {
-  @override
-  _UserTransactionsState createState() => _UserTransactionsState();
-}
-
-class _UserTransactionsState extends State<UserTransactions> {
-  final List<Transaction> _userTransaction = [
-    Transaction(
-      id: 't1',
-      title: 'New shoes',
-      amount: 60.0,
-      date: DateTime.now(),
-    ),
-    Transaction(
-      id: 't2',
-      title: 'Groceries',
-      amount: 50.0,
-      date: DateTime.now(),
-    ),
-  ];
-
-  void _addNewTransaction(String txTitle, double txAmount) {
-    final newTx = Transaction(
-      title: txTitle,
-      amount: txAmount,
-      date: DateTime.now(),
-      id: DateTime.now().toString(),
-    );
-    setState(() {
-      _userTransaction.add(newTx);
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: <Widget>[
-        NewTransaction(_addNewTransaction),
-        TransactionList(_userTransaction),
-      ],
-    );
-  }
-}
